+++ components/select/style.ts
@@ -1,24 +1,18 @@
-import { css, injectGlobal } from '../emotion';
+import { injectGlobal } from '../emotion';
 import prefixCls from '../prefix-cls';
 
 export const cls = {
   select: prefixCls('select'),
-  value: prefixCls('select-value'),
 };
 
-const cs = css`
+injectGlobal`
   .${cls.select} {
     position: relative;
     display: flex;
     margin-block-end: 8px;
-  }
-  .${cls.value} {
-    position: absolute;
-    inset-block-start: 0;
-    inset-inline-start: 14px;
-    pointer-events: none;
-    z-index: 1;
+
+    &:last-of-type {
+      margin-block-end: 0;
+    }
   }
 `;
-
-injectGlobal([cs]);