+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# [1.6.0](https://github.com/monako97/neko-ui/compare/1.5.0...1.6.0) (2023-05-09)
+# [1.7.0](https://github.com/monako97/neko-ui/compare/1.5.0...1.7.0) (2023-05-10)
 
 
 ### Bug Fixes
 
 * backtop 返回顶部无效 ([ab9c775](https://github.com/monako97/neko-ui/commit/ab9c77586be205eb7fc40245383cf36f485229da))
+* property 'content' is missing in type ([882b700](https://github.com/monako97/neko-ui/commit/882b70017ef104a278bc05de5fc546cda83290a1))
 * segmented dark style ([388f99d](https://github.com/monako97/neko-ui/commit/388f99dcefd6838ba5f39e2612c106365c5aef62))